✨CSS edits and removal of the intro slide Removes unneccecary css, fixes scrollbar bug (that just appeared out of nowhere - thanks css 🤦‍♂️) and removes the intro warning slide as it is no longer needed.
