+++ frameworks/operations/ops-leadership.md
@@ -1,5 +1,5 @@
 ---
-path: "/frameworks/operations/ops-leadership"
+path: "/operations/ops-leadership"
 title: "🎯 Ops Leadership Framework"
 sidebarTitle: "🎯 Leadership"
 sidebarGroup: "operations"
