+++ frameworks/product.md
@@ -1,5 +1,5 @@
 ---
-path: "/frameworks/product"
+path: "/product"
 title: "🚀 Product Framework"
 sidebarTitle: "🚀 Product"
 sidebarGroup:
