+++ src/pages/404.js
@@ -17,46 +17,44 @@ type Props = {
 export default class Frameworks extends React.Component<Props> {
   render() {
     return (
-      <div className="full-page">
-        <Layout data={this.props.data}>
-          <div className="strip strip--middle cover cover-md--640">
-            <div className="container">
-              <div className="grid-row grid-row--middle grid-row--center">
-                <div className="grid-col-12 grid-col-6-md">
-                  <ErrorPageHeader>404</ErrorPageHeader>
-                  <ErrorPageDescription className="u-color-grey">
-                    Sorry, we can&rsquo;t find the page you&rsquo;re
-                    looking&nbsp;for
-                  </ErrorPageDescription>
-                  <ErrorPageLink
-                    to="/frameworks/"
-                    className="o-button-text u-margin-top-large-md"
-                  >
-                    Go back home&nbsp;&rsaquo;
-                  </ErrorPageLink>
-                </div>
-                <div className="grid-col-12 grid-col-4-md">
-                  <LowFundsAnimation />
-                </div>
+      <Layout data={this.props.data}>
+        <div className="strip strip--middle cover cover-md--640">
+          <div className="container">
+            <div className="grid-row grid-row--middle grid-row--center">
+              <div className="grid-col-12 grid-col-6-md">
+                <ErrorPageHeader>404</ErrorPageHeader>
+                <ErrorPageDescription className="u-color-grey">
+                  Sorry, we can&rsquo;t find the page you&rsquo;re
+                  looking&nbsp;for
+                </ErrorPageDescription>
+                <ErrorPageLink
+                  to="/"
+                  className="o-button-text u-margin-top-large-md"
+                >
+                  Go back home&nbsp;&rsaquo;
+                </ErrorPageLink>
+              </div>
+              <div className="grid-col-12 grid-col-4-md">
+                <LowFundsAnimation />
               </div>
             </div>
           </div>
-          <div className="strip">
-            <div className="container">
-              <div className="grid-row grid-row--center">
-                <div className="grid-col-12 grid-col-6-md u-align-center-md">
-                  <h3>Did you know&hellip;</h3>
-                  <p>
-                    In 2008, there were 4,916 claims (totalling &pound;113,000)
-                    for bank notes which had been eaten by pets.
-                  </p>
-                  <h1>🐶 🐱</h1>
-                </div>
+        </div>
+        <div className="strip">
+          <div className="container">
+            <div className="grid-row grid-row--center">
+              <div className="grid-col-12 grid-col-6-md u-align-center-md">
+                <h3>Did you know&hellip;</h3>
+                <p>
+                  In 2008, there were 4,916 claims (totalling &pound;113,000)
+                  for bank notes which had been eaten by pets.
+                </p>
+                <h1>🐶 🐱</h1>
               </div>
             </div>
           </div>
-        </Layout>
-      </div>
+        </div>
+      </Layout>
     )
   }
 }
